140509_03/src/requirements.txt
fastapi==0.95.2
uvicorn>=0.29.0
pydantic>=1.10.15,<2
pyyaml>=6.0.1
numpy>=1.26.4
SQLAlchemy>=2.0.30
psycopg2-binary>=2.9.9
prometheus-client>=0.20.0
redis>=5.0.1
alembic>=1.13.2
